refactor(team-hub): #373 Phase 2 protocol.rs を機能別 module に分割
src-tauri/src/team_hub/protocol.rs (1729 行) を以下の構造に分割:

protocol/
  mod.rs          (134) — pub fn handle() entry + dispatch_tool
  consts.rs       (29)  — 定数 (RECRUIT_TIMEOUT / MAX_MESSAGE_LEN 等 10 個)
  schema.rs       (132) — tools/list の JSON Schema 定義 (11 ツール)
  permissions.rs  (45)  — caller_has_permission / builtin_role_permission
                          (Issue #136 security-critical)
  dynamic_role.rs (129) — validate_and_register_dynamic_role
  helpers.rs      (225) — resolve_targets / message_is_for_me + unit test 11 個
  tools/
    mod.rs              (29)  — pub use 集約
    recruit.rs          (355) — team_recruit (動的ロール検証 + handshake)
    dismiss.rs          (87)  — team_dismiss
    send.rs             (267) — team_send (並列 inject)
    read.rs             (67)  — team_read
    info.rs             (52)  — team_info
    diagnostics.rs      (79)  — team_diagnostics (leader/hr only)
    assign_task.rs      (176) — team_assign_task
    get_tasks.rs        (30)  — team_get_tasks
    update_task.rs      (27)  — team_update_task
    list_role_profiles.rs (50) — team_list_role_profiles
    status.rs           (14)  — team_status (旧 dispatch インライン → 関数化)

合計 1927 行 (+198 行 = doc comment / module 階層 / use 文の増加)。

公開 API:
- pub fn handle() のみ (mod.rs)。`team_hub::mod.rs:1070` の唯一の呼び出し点
  はシグネチャ無変更。pub use ではなく直接 mod.rs に置く形で維持。
- 各 helper / tool 関数は pub だが、tools/permissions/etc は private mod
  なので crate 外には漏れない。

不変式 (Issue #373 全 Phase 共通) を保持:
- IPC コマンド名・event 名 (team:recruit-request / team:role-created /
  team:dismiss-request / team:handoff / team:recruit-cancelled) は逐字保持
- JSON Schema (tool 名 / description / inputSchema) は schema.rs に逐字
  コピー (renderer / Claude Code / Codex 側 MCP クライアントが文字列マッチ
  に依存する可能性があるため)
- serde derive (TeamMessage / TeamTask / DynamicRole 等) は team_hub::mod.rs
  側にあり、本 PR では一切触らない
- エラーコード (recruit_failed / send_payload_threshold 等) と message
  文字列は逐字保持
- feature flag (legacy_message_fallback) は helpers.rs の test mod に
  そのまま移設

検証:
- cargo check --manifest-path src-tauri/Cargo.toml: green
- cargo test --lib team_hub::protocol: 11 件全 pass
- cargo clippy --no-deps: 既存ベースライン以下 (新規警告ゼロ)

Refs #373
